feat(sdk): :sparkles: 对齐 OpenAPI 类型定义，完善事件路由、测试与 payload 解码
- config.go 将 autoupdate/formatter/lsp/plugin/reference 等 any 字段替换为类型化 Union 参数
- global.go 跳过非对象 payload 的 Union 路由，避免 SSE 流中断
- global.go Phase 1 别名结构将 Payload 声明为 json.RawMessage 而非 any，
  保留 JSON.Payload 元数据的同时省去立即丢弃的 map[string]any 分配
- event_sync.go 复用 v1 事件属性类型，消除重复定义
- event.go 修正 exitCode 字段类型为 int64
- part.go 为 retry 错误 name 增加枚举类型
- 移除废弃的 V2FsService.ReadPath 方法
- 新增 GlobalEvent 元数据、SSE 路由与 124 个 payload 全覆盖测试
- 现代化测试代码，升级 gjson 至 v1.19.0

+++ event.go
@@ -3206,7 +3206,7 @@ func (r EventListResponseEventPtyExited) implementsEventListResponse() {}
 func (r EventListResponseEventPtyExited) implementsGlobalEventPayload() {}
 
 type EventListResponseEventPtyExitedProperties struct {
-	ExitCode int                                           `json:"exitCode,required"`
+	ExitCode int64                                         `json:"exitCode,required"`
 	ID       string                                        `json:"id,required"`
 	JSON     eventListResponseEventPtyExitedPropertiesJSON `json:"-"`
 }
